Create Scala 2 veneers for ModuleChoice.scala
