+++ src/main/scala/chisel3/choice/ModuleChoiceImpl.scala
@@ -2,7 +2,6 @@
 
 package chisel3.choice
 
-import scala.language.experimental.macros
 import scala.collection.immutable.ListMap
 
 import chisel3.{Data, FixedIOBaseModule, Module, SourceInfoDoc}
@@ -11,37 +10,10 @@ import chisel3.internal.{groupByIntoSeq, Builder}
 import chisel3.internal.binding.WireBinding
 import chisel3.internal.Builder.pushCommand
 import chisel3.internal.firrtl.ir.DefInstanceChoice
-import chisel3.internal.sourceinfo.InstChoiceTransform
 
-object ModuleChoice extends SourceInfoDoc {
+private[chisel3] trait ModuleChoiceImpl {
 
-  /** A wrapper method for Module instantiation based on option choices.
-    * (necessary to help Chisel track internal state).
-    * @see [[chisel3.choice.Group]] on how to declare the options for the alternatives.
-    *
-    * Example:
-    * ```
-    * val module = ModuleChoice(new DefaultModule)(Seq(
-    *   Platform.FPGA -> new FPGATarget,
-    *   Platform.ASIC -> new ASICTarget
-    * ))
-    * ```
-    *
-    * @param default the Module to instantiate if no module is specified
-    * @param choices the mapping from cases to module generators
-    *
-    * @return the input module `m` with Chisel metadata properly set
-    *
-    * @throws java.lang.IllegalArgumentException if the cases do not belong to the same option.
-    */
-  def apply[T <: Data](
-    default: => FixedIOBaseModule[T]
-  )(choices: => Seq[(Case, () => FixedIOBaseModule[T])]
-  ): T =
-    macro InstChoiceTransform.apply[T]
-
-  /** @group SourceInfoTransformMacro */
-  def do_apply[T <: Data](
+  protected def _applyImpl[T <: Data](
     default: => FixedIOBaseModule[T],
     choices: Seq[(Case, () => FixedIOBaseModule[T])]
   )(
